AUDIT-57:Moved the getAuditedEntitiesNames from UtilClass to Service one and make it secure with auth privilege
